blog/techy/requirements.txt
fastapi==0.60.1
uvicorn==0.11.8
gunicorn==20.0.4
SQLAlchemy==1.3.18
bcrypt==3.1.7
PyJWT==1.7.1
python-multipart==0.0.5